requirements.txt
Flask==2.1.1
tensorflow==2.19.0
Pillow==9.1.0
numpy==1.26.0
Werkzeug==2.0.1
